server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic[email]==2.8.2
httpx==0.27.0
PyJWT==2.8.0
requests==2.31.0
psycopg2-binary==2.9.10
supabase==2.4.2
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.9
google-auth==2.23.3
resend==0.7.0